Armada documentation via build_sphinx
This PS adds tooling and automation to automatically generate
Shipyard's documentation into feature-rich HTML pages that can
be hosted.

To run the documentation job, simply execute:

    tox -e docs

warning-is-error = 1 was set under [build_sphinx] in setup.cfg
so that warnings could be caught.

A number of warnings/errors while building the docs are have been
corrected by this PS as well.

Change-Id: I445369c802e559c629a685024575bd562ea05a1b