bug(tests): Unskip Armada API tests
This PS sets the foundation for unskipping Armada API unit tests
by making necessary framework changes where necessary as well
as test refactoring so that the tests not only can be unskipped
but also execute successfully.

However, until a feature-rich testing framework is in place, it's
not possible to do end-to-end testing as mocking is currently used
to stub out Tiller API.

Additional unit tests in the same vein, like negative RBAC tests,
will be done in follow-up patches.

Included in this PS:
  - Added Policy fixture for asserting that expected policies are
    enforced
  - Added unit tests for tiller/armada API
  - Added test_utils module for providing utility helper functions
    for tests
  - Modified default value of ``[DEFAULT] middleware`` to ``True``
    as the `middleware` opt is a `BoolOpt`, not a `StringOpt`.

Closes #184

Change-Id: I2f454b27b014875bc35fd706f7c0d05364ce562a