fix(tox): Add whitelist externals to tox.ini to avoid warnings
This PS adds whitelist externals to Armada's tox.ini to avoid
warnings that are issued while running tox -e py35 or tox -e pep8.
Example:

WARNING:test command found but not installed in testenv
  cmd: /usr/bin/find
  env: /opt/stack/armada/.tox/py35

This PS also resolves a critical error while running
tox -e coverage, because tox defaults to python2.7 for
the coverage job, and Armada's hapi module is not compatible
with python2.7. The fix involves setting the testenv's
basepython to python3.5.

Change-Id: Ice0074c3453c80290886f27a4518fbed01722572