fix(chartbuilder): Fix ChartBuilder failing on get_files()
This PS fixes a recent commit to chartbuilder module [0]
causing HelmChartBuildException and DependencyException
errors to be thrown. This is because of the recent implementation
of get_files() which uses string primitives rather than instances of
`google.protobuf.Any` [1].

The solution was to wrap each file name in the above instance.

Unit tests were added to validate much of the ChartBuilder class.

Change-Id: Iddb276b917b43754511234737b749182caf4ff14