tests(git): Unskip git tests and add conditional skip test.
This PS unskips the last remaining git test in test_source which
had an unconditional unittests.skip() decorator applied to it.
The underlying reason why it was skipped was fixed: calling
source_remove without an existing repository as the path raises
the expected exception.

This also adds a conditional skip to the git clone unit tests
in the event that the user doesn't have network connectivity.

Change-Id: Ic300c48a4e0c361e37ec15cd3675415f8170a817