bug(manifest): Allow specific manifest to be specified
This PS allows users to specify the manifest file to use
by the Armada handler by introducing a new flag called
`target_manifest`. This flag was added to the API and
CLI.

A foundation of unit tests for the manifest handler
are included in this PS. Most of the coverage is aimed
at checking the various success and failure cases]
surrounding the new target_manifest feature.

Also updates documentation to convey information about
the new flag.

Change-Id: I1d5a3ecc1e99b6479438d0ee5490610178be34fe