Set the API port from chart env var
For the deployment of armada, the entrypoint.sh needs to have
the exposed uwsgi port matching the deployment of the api pod.
This change sets the value in the pod.env.armada_api seciton
so that the environment variable is set up by the chart rather
than relying upon the default value set in the entrypoint.sh
script.

Change-Id: I7535abac7c252f5c65447476e02f8eb54647e966