[WIP] Updated Makefile
ARMADA-29: CI/CD Armada Testing Framework Entrypoints defined

It does not appear the the docker run command to call tox tests/unit
is necessary.  The Makefile already includes several testing checks.

Added a few additional targets as specified in the upc-integration
code-convenstions to satisfy the clcp Makefile standards.

As far as I am aware, all that is left is adding the docker image to
the run_armada file. I am currently working to track down the docker
image for armada.

Change-Id: Ia920a856ed6d607fcf38d5f85dbb9a0c8aae4fa6