[test] Increase armada.handlers.armada test coverage to 59%
This is a multi-part PS because these patches may include small
fix-ups to the code base itself, so the intention is to keep
the patches small and easily reversible. This patchset introduces
the following:

  * html coverage report (execute tox -e cover then open index.html
    under htmlcov folder which is created by py.test)
  * adds unit tests for pre_flight_ops for negative cases
  * additionally, the previous check was `if known_releases is None`
    but this check isn't correct since
    `known_releases = self.tiller.list_charts()` returns a list.

Removes the following:

  * removes raising an exception if no known releases are found
    because this exception isn't needed; it leads to unnecessary
    complexity

There are still multiple scenarios and edge cases that need to be
tested like charts with pre- and post-actions, upgrade sections,
timeouts and so forth. Those scenarios will be covered in
follow-ups.

Change-Id: Ib29d7d8d0c3b686a36c5a87fc46d4594bb1838a6