Works around low RELEASE_LIMIT by increasing it
We have been deploying in the low 60's of charts in many environments,
but the current treasuremap environment is running 68.

This led to a situation where tiller would find all 68 charts, but then
return only the newest 64 (RELEASE_LIMIT) in its response.  That meant
that armada did not see those charts when deciding whether to install or
upgrade.

Change-Id: I5d2a06f806006947ad48cc0e24b3f91baa3f37fd